test: run the guard's behavioural suite once, and only its sentinel everywhere
The continuous-integration .NET job reddened on the previous head at
StatusPaletteTests.BrushFor_ANameTheFoldNeverEmits_IsTheBrokenSentinel_NeverGrey
with "Call from invalid thread". That test reads an Avalonia styled property,
which throws unless it runs on the dispatcher thread, so it is order and
thread dependent. It passes here in isolation and in a full run of its own
assembly, main has been green for twenty-five consecutive runs, and the three
previous heads of this branch passed with this guard already linked in.

I have NOT traced it to a cause and am not claiming one. What is true is that
the only change at that head was to this guard's own tests, and those tests
were being compiled into all seven assemblies - building real git
repositories and launching real processes in every one of them, including the
assembly with the thread-sensitive tests.

That was poor design independently of whether it caused anything. The
behavioural suite does not need to run seven times; running it there buys no
coverage and changes the conditions other suites run under. A mechanism has
no business doing that, and the cheapest way to stop being a suspect is to
stop being present.

So the guard still goes into every test assembly, because that is the
mechanism, and a small sentinel goes with it, because each assembly must
prove its own module initializer woke up - one that silently does not run
leaves no trace and every other test still passes. The behavioural suite now
compiles into Core.Tests alone.

Avalonia.Tests goes from 299 tests in 11 seconds to 248 in 1.
